APIs/azure.com/network-ipGroups/2019-09-01/python-aiohttp-client/tests/test_ip_groups_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.ip_group import IpGroup
from openapi_server.models.ip_group_list_result import IpGroupListResult
from openapi_server.models.ip_groups_list_default_response import IpGroupsListDefaultResponse
from openapi_server.models.ip_groups_update_groups_request import IpGroupsUpdateGroupsRequest


pytestmark = pytest.mark.asyncio

async def test_ip_groups_create_or_update(client):
    """Test case for ip_groups_create_or_update

    
    """
    parameters = {"etag":"etag","properties":{"firewalls":[{"id":"id"},{"id":"id"}],"ipAddresses":["ipAddresses","ipAddresses"],"provisioningState":"Succeeded"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/ipGroups/{ip_groups_name}'.format(resource_group_name='resource_group_name_example', ip_groups_name='ip_groups_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_ip_groups_delete(client):
    """Test case for ip_groups_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/ipGroups/{ip_groups_name}'.format(resource_group_name='resource_group_name_example', ip_groups_name='ip_groups_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_ip_groups_get(client):
    """Test case for ip_groups_get

    
    """
    params = [('api-version', 'api_version_example'),
                    ('$expand', 'expand_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/ipGroups/{ip_groups_name}'.format(resource_group_name='resource_group_name_example', ip_groups_name='ip_groups_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_ip_groups_list(client):
    """Test case for ip_groups_list

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Network/ipGroups'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_ip_groups_list_by_resource_group(client):
    """Test case for ip_groups_list_by_resource_group

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/ipGroups'.format(resource_group_name='resource_group_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_ip_groups_update_groups(client):
    """Test case for ip_groups_update_groups

    
    """
    parameters = openapi_server.IpGroupsUpdateGroupsRequest()
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PATCH',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/ipGroups/{ip_groups_name}'.format(resource_group_name='resource_group_name_example', ip_groups_name='ip_groups_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

